requirements.txt
colorama==0.4.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
